files-view: Restore search state on view switch
As follow-up to not treating search as a navigation node[0], I've
removed some intrincate logic to handle the case where we go back or
forward to a search directory.[1]

Except... this logic was also needed for the case where we reload a
search directory in a new view during view mode (grid/list) change.

As a consequence, switching view mode during search results in an
empty results view, even if there were some before.

So, let's restore the important bits that were removed in [0].

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3114

[0] commit d06b4d6a81d52dc085fba33a6c91d479841228c9
[1] commit 570e5c59c372058b4af4c754ad954832f2a5bbfe