files-view: Restore searched location state
When switching view mode, a new view is created. If we are in a search,
the new view has no previous location to go back to when canelling
search, so it falls back to the user home directory.

In order to improve this situation, let's have the new view assume that
the queried location is the original location.

(This fails for global search, but the proper fix requires larger
changes, which are not suited for a bugfix in the stable branch.)