file-operations: simplify link name creation
Naming additional links to a file "Another link to" or "Xth link to" is
gimmicky and results in a lot of translation overhead. Instead only
append a counter as a suffix, keeping the file name predictable and
simplifying the code.

The new code makes use of `nautilus_filename_shorten_base` and is
placed in filename-utilities, usages are adjusted to it.