file-operations: shorten filenames by base
`nautilus_filename_shorten_base` shortens a file name to fit within a certain
limit, by trimming the base part of the filename. It can thus replace
`shorten_utf8_string`, as that was only used for file names.

The method is placed in a new 'filename-utilities' file to allow creating a
unittest for it, as well as in preparation for moving other file name functions
there.