file-operations: shorten filenames by base
`nautilus_filename_shorten_base` shortens a file name to fit within a
certain limit, by trimming the base part of the filename. It can thus
be used to shorten file name without mangling extensions and other
suffixes. It gets used in two places where a file operation can lead to
a file name conflict and a suffix is added to resolve it.

The method is placed in a new 'filename-utilities' file to allow creating a
unittest for it, as well as in preparation for moving other file name functions
there.