file-operations: turn self-check into proper unittests
Turn the file duplication testing code into a dedicated unittest.
Unittests are more straightforward to run than the eel self-check
mechanism. This also allows dropping eel dependencies and reduces
the size of the file-operations file.

The concrete duplication syntax is abstracted away with a macro for
better readability of the comparisons.