appdata: Remove mailing list, add discourse link
The mailing list is gone and the replacement is discourse.

As the update_contact tag expects an email address, remove it without
replacement. This tag is meant only for distributors[0], and our
distributors know well how to contact the GNOME project. Also, problems
with this metainfo can just be reported as bugs, so the bugtracker url
tag above is the obvious route.

The discourse is added as a new tag, meant for users to get in contact
with our community for any other reasons.

[0] https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-update_contact