file-changes-queue: Queue unmounts specifically
The window slot is using a complex logic to know when the current
location is gone due to being unmounted, and it works only for the
root location of the mount. Ideally we should do the same if the
current location is anywhere in the unmounted hierarchy.

NautilusMonitor gets about unmount events for all native files,
and forwards this through this pipeline, resulting in the file
being marked as gone and the NautilusFile::changed being emitted.

-monitor -> -file-changes-queue -> -directory-notify -> -file

So, we need to additionally channel the information that the file
was unmounted through this pipeline.

As a first step, introduce a new path to queue unmount events. For
now, it just conflates then with other removals, but this lays the
ground for behavior changes next commit.