compress-dialog-controller: add password confirmation entry
As described in #3128, the compress-dialog currently does not ask the user
to confirm the password when creating an encrypted archive. This could lead
to typos in the password and potentially data loss if the original files are
deleted after compression.

This MR adds a password confirmation entry to the compress-dialog, preventing typos
and matching the example password entry from the GNOME HIG.

Closes #3128