file-operations: simplify file copy name creation
The mechanism with which file names for copies of existing files are
created is very convoluted and not following good practices for easy
translatability. In practice having a "(4th copy)" does not add much
benefit for the user.

Instead only append a simple string with an optional counter as a
suffix, keeping the file name predictable and simplifying the code.

The new code is placed in filename-utilities and usages as well
as tests are adjusted to it.

Fixes #3151