file-operations: simplify file copy name creation
The mechanism with which file names for copies of existing files are
created is very convoluted and does not follow good practices for easy
translatability. In practice having a "(4th copy)" does not add much
benefit for the user.

Instead only append a simple string with an optional counter as a
suffix, keeping the file name predictable and simplifying the code.
Essentially this reverts f3ce0116eb3d2c905a6b6132445767cf2260ea45 in
that regard.

The new code makes use of `nautilus_filename_shorten_base` and is
placed in filename-utilities, usages are adjusted to it.

Fixes #3151