window: make Alt-Down behavior symmetric to Alt-Up
Previously Alt-Down was an alternative to Enter. It opened folders and
files. All other Alt-[arrow key] combinations only navigate, so opening
of files was strange behavior for a navigation shortcut. Similarly,
Alt-Down and Alt-Up did not form a symmetric pair, like Alt-Left and
Alt-Right.

So to create the symmetry, change Alt-Down to navigate into previously
opened folders, if those folders reside in the current location. This
allows for a new way of easily navigating up to a parent folder with
Alt-Up and quickly navigating back to the previous location with
Alt-Down, without risking to overshoot beyond the desired location,
as e.g. navigation with Alt-Left can cause.