Revert "properties: Use cache nautilus file when opening from sidebar"
This reverts commit 5d371babaf464db61c1134568a72d4f553680439.

The result is the same, because `nautilus_file_get_by_uri()` would just
return a new reference to the already existing file.

Reverting because, with the next commit, we should not be relying on
the file to be set on the row.


(cherry picked from commit 340507b25a24e3e47c898f7dd45d193ebd685925)