gtksidebarrow: Don't keep ref to gone file
The row holds a NautilusFile instance to cache information needed for
Properties dialog and DND. [0]

But not listening for changes means we never release the reference when
the file is marked as gone. This naturally leads to bugs.

Listen "changed", clear when gone, and reconnect on-demand, but even
then protect against ref'ing a gone file (which shouldn't happen but
still does for self-owned files).

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2922

[0] dc2711bd74056ed1711c0bce8dfcd9d0056bc3ad