window-slot: Release viewed file as soon as it's gone
Anyone keeping a reference to a NautilusFile instance should listen for
changes and release the reference when the file is marked as gone.

The slot is listening to the ::changed signal and, if the file is gone,
it changes location, which, as a side effect, releases its reference.

However, this side effect happens only after info for the new file is
ready, which may not happen until a future iteration of the main loop:

-> viewed_file_changed_callback()
 -> nautilus_window_slot_open_location_full()
  -> begin_location_change()
   -> nautilus_file_call_when_ready()
    -> (ASYNC CALBACK) got_file_info_for_view_selection_callback()
     -> setup_view()
      -> change_view()
       -> nautilus_window_slot_update_for_new_location()
        -> nautilus_window_slot_set_viewed_file()
         -> nautilus_file_unref() && g_signal_handlers_disconnect_by_func()

As the file release happens asynchronously, if we get a second ::changed
emission for the same gone file in the current iteration, we setup double
location change and double async side effects, which inevitably leads to
a buggy state which is prone to crash. (This may happen currently if
nautilus-monitor.c report an unmount twice, which happens with GOA mounts
which have a shadow mount with identical root URI.)

This async unref has been happening at least since commit 15d056a712cf3f6935bcb95003881aea01475693
It removed cancel_viewed_file_changed_callback() from begin_loading(),
which had been added by 3ca224bdc680b77a364985c8924126c0a87b542c

-> viewed_file_changed_callback()
 -> nautilus_window_slot_open_location_full()
  -> begin_location_change()
   -> cancel_viewed_file_changed_callback()
    -> g_signal_handlers_disconnect_by_func()

But even then, it didn't unref the file synchronously, it just avoided
the unwanted double ::changed notification. Nothing good comes from
keeping a gone file around for the next few main loop iterations.

In fact, back when this whole machinery was written, we had a spatial
mode, which opened each file in its own window. A corolary of this was
that, if the file was gone, then the window should be destroyed too.
That was the case when cancel_viewed_file_changed_callback() was first
introduced by commit fcb7d41216b32d39072eda0d47c17d5dccc482da

-> viewed_file_changed_callback()
 -> nautilus_window_destroy()
  -> nautilus_file_unref()

So, back to present, we should once again make sure we both disconnect
and unreference the viewed file as soon as we get notified it's gone:

-> viewed_file_changed_callback()
 -> nautilus_window_slot_set_viewed_file()
  -> nautilus_file_unref() && g_signal_handlers_disconnect_by_func()

Actually fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2922 for good.


(cherry picked from commit 2cd08a4c9202d18396e4e06ab79f6e920e1e4bb1)