file-name-widget-controller: Unify name_is_valid
In preparation to stop deriving from this class, stop overriding this.

"Folder" or "File" strings can be picked relying on the new property.
Custom strings for archives are dropped; archives are files too and
some errors would never apply anyway (an archive name never ends with
a space, and it's never simply "." nor "..", because it has an archive
type filename extension appended).

(If necessary we could reintroduce the archive strings later by turning
boolean :target-is-folder property into an enum property.)