file-name-widget-controller: Drop ::cancelled signal
Instead of a "cancel" signal, always invoke the callback but with a
NULL name in case of cancellation.

This is a temporary measure to stop deriving from this class, because,
for the time being, the view still needs to keep a reference on the
controllers, and this reference needs to be cleared on cancellation.