file-operations: Unify dialog creation methods
In preparation for using `AdwAlertDialog` instead of `GtkDialog` for
operation dialgos, use a default `GtkMessageType` for the dialogs, as it
does not make a difference anyway.

This allows to remove the now unused `message_type` and replace the
equivalent `run_question()`, `run_warning()` and `run_error()` with a
unified `run_dialog()`.