compress-dialog-controller: Actually cache passphrase
The `update_passphrase()` function has been introduced to avoid
duplicating code for both passphrase fields.

However, it's passed the value of the passphrase field in the self
struct, not the address of the field. As such, the field value remains
unchanged, except it now points to memory that's been freed. Oh no!

As such, the Create button is never sensitive. It's still possible
to force-create the archive by activating the name entry, but the
resulting ZIP archive is not encrypted at all.

Let's actually pass the address of the cache we want to update.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3193