compress-dialog: Ignore initial format setting signal
When setting up the dropdown, we set its initial value to that of
the last used format, stored by the GSettings backend.

This results in the `update_selected_item()` callback being called.
This callback does three things:

  1. update visibility of passphrase widgets;
  2. store the new format in the GSettings backend;
  3. trigger filename validation because extension may have changed.

While the 1st effect is desired during the initial setup of the dialog,
the other two are undesired: storing the just-read settings value is
futile, and triggering filename validation at this time, before the
containing directory is passed to the validator, is wrong and triggers
NautilusDirectory warnings.

So, factor out the first effect and block the signal callback in the
setup phase.