compress-dialog: Stop caching passphrases
The ::changed signal handlers copy the text from the child widget into
our instance struct, which is pointless now that the passphrase
matching is done in an closure through expression binding.

Furthermore, we have been leaking the cached confirmation passphrase.

Let's drop this the caching and get the passphrase at the time of use.

Related to https://gitlab.gnome.org/GNOME/nautilus/-/issues/3193