compress-dialog: Don't bypass passphrase entries
Activating the name entry (e.g. pressing Enter while it's focused)
triggers the creation of the archive with that name, if if valid.

But this is undersirable when passphrase is expected to be filled
and to match the confirmation passphrase. In such case the "Create"
button is insensitive as it should, but the entry activation still
goes through.

So, if passphrase is required, let's focus the passphrase entry
instead of creating the file.