window: Make tab bar respect gtk decoration layout setting
When changing the placement of titlebar buttons from right to left in
Tweaks, the close button of the tabs would stay on the right side
instead of moving to the left like the titlebar buttons.
Epiphany fixed this by connecting to the "notify::gtk-decoration-layout"
signal and inverting the tab bar layout if necessary.
I mostly just copied the code from Epiphany.

Fixes: #3137