main: Recommend admin backend if run as root
GUI apps shouldn't run as root, neither directly nor through sudo.
Explaining why that's a very bad idea for the 1000th time is out of
scope for this commit message.

Introduce a small nudge to guide people to use the admin backend.

In order to have it localized, initialize the locale here. Also
include <locale.h> unconditionally. (We already do so elsewhere.)