files-view: Drop now-useless processing step
The whole purpose of `process_new_files()` was moving non-ready files
into the "waiting room", passing the rest to `process_old_files()`.
The "old files" naming was a misnomer; "ready files" would have been
a better fit better. But now that we don't care whether they are ready
or not, let's just call them "pending files".

This completes the job of the previous commit in reverting the logic
introduced by all the way back in 2001.[0]

[0] https://gitlab.gnome.org/GNOME/nautilus/-/commit/5d691d775097c9d73f4e544ec36c58bec1c05b4c