file: Use placeholder while loading thumbnail
Now that we add files to the view before their thumbnail is loaded[0],
We see their MIME type icon before the thumbnail appears.

This is inconsistent with what happens when the thumbnail is generated.
In that case, we show a placeholder icon which doubles as a loading
indicator: image-loading.

Let's do the same thing when an existing thumbnail is not yet loaded.

[0] commit a1e97e846e7a659720962feb26f540f228b2a5f0