view-model: Use binary search to find index
We've been using a linear search algorithm to find the position of an
item in the model. Unsurprisingly, this is a performance liability. So,
we've been avoiding it in general and chiefly in loops.[0]

But I just occurred me that the list model we are iterating is a sorted
list model. So, we might as well use a binary search algorithm to make
it O(logN) instead of O(N).

The results, in my machine, on a folder with 1757 items, are obvious:

Index         Before         After
    0           0 ms         13 ms
    2           2 ms         13 ms
   21           7 ms         15 ms
   36          12 ms         14 ms
   56          18 ms         13 ms
  121          35 ms         14 ms
  161          51 ms         14 ms
  271          84 ms         15 ms
  311         114 ms         15 ms
  756         223 ms         17 ms
 1251         315 ms         15 ms
 1746         466 ms         14 ms
 1757         471 ms         15 ms

Also rename get_index() to find(), because it's not a getter but more
like g_list_store_find().

[0] commit 5b29b99b813d21ee4bbd23157c6cb2fe0055331d