files-view: Set selected files even more efficiently
The previous optimization was very modest.[0] While it did lower the
processing time in common cases, the time complexity was still O(N×M).

Now that the model finds index at logarithmic time, we can revert most
of that change and further optimize the previous solution. Knowing that
the list of items to find and the full model are sorted the same way,
we can perform multiple binary searches on increasingly smaller ranges.

This makes the whole thing O(logN × M) at worst.

[0] Commit 86cb3f1e97d74f863b4239b58548162f035608f8