properties-window: Offer to set custom icons on folders only
Custom icons on files are easily confused with thumbnails and
might even be used to misrepresent a file for no good reasons.

Now that the feature is more visibly discoverable, it may also be
a distraction, so let's present it only where the relevant use cases
are: folders.

The ability to remove custom icons from regular files is kept,
to avoid mistaking them for thumbnails.