progress-indicator: Give feedback for quick operations
When a long-running (>=2 seconds) operation completes, we show a toast
in the active window, with possibly a button to open the destination.

For short operations the only feedback we have is when the operation
adds items from the folder that's currently being viewed, in which
case those added items are selected and scrolled into (if needed).

But if the destination isn't being viewed, and the operation completes
in less than 2 seconds, there's no visual feedback that the operation
has happened.

This happens because we are only connecting to "finished" signal on
operations which we estimate to take longer than 2 seconds. Indeed,
every time we update the progress indicator/operations popover, we
disconnect and reconnect signals.

Let's connect the signals only once, when the operation info is added,
and let's do it regardless of its estimated duration. Instead, let's
check how long the operation took before showing the notification;
this allows the window to decide not to show any notification if the
operation was short and the destination is being viewed, which is
the case where we already have adequate visual feedback and a toast
would only add noise.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/267