Revert "files-view: Don't update sort metadata if no change"
This reverts commit a6bcce3d9a2a678313051c86e633d574fc345a5c.

The reverted commit was intended to mirror the behavior when the sort
metadata was in list-base.  In files-view old_value will always equal
value because we are listening to the action-state-change signal of the
ActionGroup, which is emitted *after* the value is changed.
In list-base we were using the change_state callback of the
ActionEntry which was in charge of actually changing the value.

It's safe enough to go ahead and set the metadata. Worst case
scenario we set it in a redundant manner.