file: Batch rename rework
Move conflict resolution to nautilus-file.

Use the following rough algorithm to resolve conflicts:
1- If file is the start of a chain, start processing it until the
end of the chain. If it's not the start of a chain jump to step 3.
The file is the begining of a chain if the targeted new name
doesn't conflict with another file to be renamed, but may conflict
with a name that is designated as stub.
2- Try to rename the file at the start of the chain, if can't be
renamed (due to a conflict or depending on a stub), mark it as a
stub. Remove the file from the chain and go to step 1 with the
next file in the chain.
3- Follow the chain until: you find the beginning on the of the
chain, or you reach the same file you started with. If you reached
the file you started with, then it's cycle. break the cycle with a
temporary name, then go to step 1 with the file at the start of
the chain.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/1443