properties-window: Reuse GFiles and store them in a hashtable
When displaying the properties of a large number of files that are
not nested, an inefficient code cause the dialog to take 100+
seconds to open. Instead of generating the entire list again for
every check, store them in hash table before iterating, and search
using the hash table. This provides extreme time saving factor
with large file count.