files-view: Manage |view.sort
Nowadays, the action and its possible states are exactly the same for
both list and grid views, with even the sort menu being shared.

But each view still implements the action separately, which requires
acting on the NautilusFilesView-owned view group. This is fine as long
as they subclass it, but we want to make it a child widget instead of
a subclass.

So, manage the action upclass and bind its state to the sort state of
the internal views, through a new NautilusListBase property. This
should provide flexibility in the future to rebind from one internal
view to another.