issue_templates: Don't discourage filing issues
Based on feedback by @mcatanzaro on #gnome-design, the current issue templates may discourage filing issue templates when the bug is already fixed in main, even if it's affecting the supported stable versions.

What we want to know is whether it's happening in main or not.
Also, in order to use consistent language across the template text, converge on "latest development version".