file: Allow reading filesystem info for files
There's no reason for this to be restricted to directories. it was
introduced in [1] which seems to just copy the logic from [2] which
was meant for `filesystem::type`. I can't find any commit or
documentation change that states it was ever exclusive to
directories though. Also protect the reading with a check to wether
the filesystem metadata is up to date.

[1] 99f55748ef69a68df86b9a039998ee576d59b898
[2] bfd9ccf65c1b85dd6d38e17410debc23cb3c7daf