Revert "view-model: Use binary search to find index"
This reverts commit 499cb1fc687a3ce6b0306bb51d233c57d588c6b0.
But keeps the function symbol rename.

The binary search requires full sorting, but our sorting
algorithm allows partial sorting. The benefits of the
optimizatio are not worth the trouble.

Instead of trying to search faster, we are going to the other
route: trying to search less (or not at all).