window-slot: Reconnect all handlers upon navigating to an invalid location
When navigating to an invalid location (i.e. through the location
entry, a bad bookmark, etc), the file add/changed handlers are no
longer present in the view until another location change (or reload)
happens.  The existing call to nautilus_window_slot_set_viewed_file()
is not enough to fully reestablish the handlers which were removed from
the call to nautilus_window_slot_stop_loading() (which
ultimately calls disconnect_directory_handlers() in files-view).

There's an explicit comment introduced at [1] which makes it clear
that we want to remove these handlers early for performance and search
reasons. Otherwise we could have delayed removing the handlers which
would have allowed us to not call nautilus_window_slow_force_reload() at
all.  But since this invalid location is just a corner case which
most likely won't happen, let's keep the typical case optimized.

Alternatively, we could create a new function in files-view to
only reconnect the handlers but that too seems to open up a can of
worms for little return.

This bug was introduced in [1]

[1] 9e1e14db2f4c5e273f6e37222f3bac49fa64dd8c