list-base: Stop implementing .scroll-to-file()
NautilusFilesView can just call the set_cursor() method instead.

There is a minor behavior change: now it sets focus in addition
to scrolling to the target file. It looks like a good change,
actually, because otherwise we would scroll to the position but
upon arrow key press jump to the start of the list...