files-view: Minor selection optimizations
Mostly avoiding overhead of g_list_length() when we only need to know
whether the selection is empty or contains a single item. But also in a
few cases avoid allocating a file list at all by checking the selection
bitset directly or doing it only if needed (like the debug case).

While touching it, also unexport a function which is no longer used by
subclasses and remove its obsolete documentation comment.