files-view: Drop .reveal_selection()
With one exception, it's called right after setting selection, so we
can simply do both things at the same time with set_cursor().

The exception is revealing the item for the popover to attach into,
for rectangle computation, which is easily solved by having the
reveal (i.e. scrolling) be done by the rectangle computing function.

This is one less virtual function for NautilusListBase to inherit,
as part of the plan from https://gitlab.gnome.org/GNOME/nautilus/-/issues/3042