files-view: Drop .get_first_visible_file()
Let's admit it: it's utterly broken.[0]

It relies on intensive fragile computation (shared with thumbnail
priorizatio), because GTK doesn't give us the visible range in
item positions.[1] There is no hope to fix this without GTK help.

This is only used by the slot when switching view mode anyway, and
only if nothing is selected. The cases where this is most benefitial
are very long lists, but these take long to load and the scroll
happens at the end mostly as a surprise.

This is one less virtual method for NautilusListBase to inherit.

[0] https://gitlab.gnome.org/GNOME/nautilus/-/issues/3034
[1] https://gitlab.gnome.org/GNOME/gtk/-/issues/4688