window-slot: Always select file to activate
The `.pending_scroll_to` file is only ever relevant if we don't set a
`.pending_selection`. This only ever happens when we are asked to
handle e a file which extracts.

In that case, there is no reason not to additionally select it too.
The extracted files are going to be selected afterwards anyway.

This allows us remove some now-dead code. There is no optimization
here though, because nautilus_files_view_call_set_selection() runs
a linear search not unlike nautilus_view_model_find().