view-model: Drop find()
This wraps a linear search in an apparently simple call, which invites
lazy coding and invites future performance regressions.

This is not to say there is no need to iterate the model linearly. But
when doing so, it's best to do it explicitly, in a for/while loop, to
make it easy to spot performance issues.