grid-view: Workaround weakref bug
When destroying the grid view, we get these warnings:

(org.gnome.Nautilus:7844): GLib-GObject-CRITICAL **: 13:38:04.860: g_object_weak_unref: couldn't find weak ref 0x7f52797428b0((nil))

This is due to the destroying the listitems on which the same
expression binding has set 2 weak refs (as _this and as object).
The weak ref on '_this' tries to remove the weak ref on 'object'
through gtk_expression_unwatch(). But, if the later weak ref has
already been triggered, there is no weak ref to remove anymore,
hence the warning above.

As a workaround, call gtk_expression_unwatch() on cell destruction,
which it warrantied to happen before listitem is destroyed, because
the listitem owns the cell. This way, the weak refs are cleanly
removed while the object is still alive.

For some reason, the row factory on list view doesn't seem to
trigger these warnings. No clue why.