files-view: Don't warn if requested selection doesn't exist
More often than not, there is a legitimate reason for the
requested selection not to be found.

E.g.: You are on list view with expandable folders enabled,
and a file in an expanded subfolder is selected. When you
switch to grid view, it tries to restore the previously
selected item, but it's not found anywhere anymore.

The  next commit is going to introduce another such case.