file: Don't update files in info query async callback
The function used to update NautilusFile internal information after
being renamed are called in an async, which causes havok since the
batch renaming requires strict ordering to not add a NautiluFile
to a NautilusDirectory that has the same name. Instead, collect the
files that are changed while renaming and invalidate them in idle
callbacks.