window-slot: Switch view mode without force reload
Reloading the directory can be slow and resource-intensive.

Now that the actual grid and list views is not the same object as the
one which loads/monitors the directory (i.e. NautilusFilesView), we
can swap the inner view while keeping the loaded directory unchanged
and the model mostly unchanged.

(I say "mostly" because we need to collapse expanded rows before
switching to grid view, which would otherwise display subfolder
contents as if part of the parent folder.)

Resolves https://gitlab.gnome.org/GNOME/nautilus/-/issues/3042