list-base: Make model a RW property
We can't keep using nautilus_files_view_get_model() if we stop being a
subclass of NautilusFilesView. Additionally, if we want to change view
mode while keeping the model around, we must be able to set a model on
the view after creation. This is one of the final pieces for the plan
detailed in https://gitlab.gnome.org/GNOME/nautilus/-/issues/3042

This change implies we need to cope with the model-less state, i.e.,
guard against model being NULL. Additionally, for list view, it means
the sorter and the expand_as_a_tree flag must be set as soon as the
model gets set the model.