window-slot: Prioritize ancestor selection when going back
The last commit introduced restoration of selection from history.

When going back, we want to select the parent (or other ancestor),
which is usually, but not always, the saved selection. Not to change
the pre-existant behavior, ensure the ancestor selection always
takes priority over the requested selection when going back.