network-address-bar: Recycle widgetry
The new design for Network view uses an address entry bar which is
similar to the one from Other Locations view. Instead of reinventing
the wheel, let's salvage some code.

In addition to formatting the code to nautilus style, this also makes
a couple of visual changes (a label and an icon) to match the mockups.

https://gitlab.gnome.org/GNOME/nautilus/-/issues/2785