network-directory: Provide recent servers
These used to be displayed in a popover attached to the address entry
in the old Other Locations view.

The new design asks for them to be provided in-line with other items
in the view.

As such, reuse and expand the recent servers code to produce GFileInfo
models for the NautilusNetworkDirectory to create new virtual files
from. Unlike the virtual files aggregated sourced from the network:///
and computer:/// GVFS backends, these virtual NautilusFiles have no
corresponding GVFS presence, and are entirely "nautilus-land" files.

Big credit goes to the people who designed the nautilus directory and
file abstractions with this in mind.