list-view: Guard against gone file usage
Between the time this was scheduled and the actual call, the file
may have been removed from the directory and, as such, from the
view, and the only reason the row, item and file weren't destroyed
is because we kept a reference to the row. But it's invalid to
access the file now.