window: Change issue-url for "Report An Issue"
"Report An Issue" redirects to https://gitlab.gnome.org/GNOME/nautilus/-/issues/new directly.
However, the users that are not logged in, would be greeted by a login screen.

This fix would redirect all users to see all the existing issues, irrespective of whether they are logged in or not.

Fixes: #3220