files-view: Update file mount info when mounted
Otherwise, nautilus_file_can_unmount() will still return FALSE for
a file which is already mounted, until the can-unmount attribute
is updated asynchronously by the file monitor.

This is needed for the next commit to be able to check whether
unmount is possible as soon as the mount operation completes.