progress-indicator: Remove operations button
The operations button is going away and will be replaced by a short
sidebar list in the following commits.  The sidebar button is now
useless, as well as the associated revealer (which comes included
in the AdwToolbarView). We can also drop a bunch of animations
that are no longer necessary.  Those animations were originally
necessary to get the attention of the user (since the operations
button was small and somewhat inconsplicuous).