progress-indicator: Stop rea-dding all progress infos
The progress indicator uses a ListModel for the ListBox, but rather,
than updating, the ListModel, it constantly dumps its model and
re-reads all of the progress infos from the progress manager upon
any change. This is wasteful and messy.

There are a few main components here:
1) The filtering.  We retrieve the entire list of possible infos
from the progress-manager, and subsequently filter that list by
calling should_show_progress_info().  We can instead handle this
by having the info itself determine if it should be shown.
2) Sharing.  All of the progress indicators share a common list
with a singleton progress manager.  This is unnecessarily complicated
and leaves room for error. In subsequent commits we'll have the
progress indicator take ownership of progress infos, so we
don't have to worry about a single shared list.