progress-indicator: Use filtermodel instead
Now the progress indicator will own its progress infos. We can use
the new properties to filter out only long running items. The bool
filter needs to be notified when its expression changed, so simply
call filter_changed when that may have happened.

Here we also include the fix for what previously was handle by
"progress manager viewers".  If a popover is visible, we don't
remove the info from the model, since the user might want to study
the info while the popover is actively open. Importantly, now that
each progress indicator owns its progress infos, this means that
other progress indicators can freely let go of an info which may
persist in the progress indicator with an active popover.