progress-info-manager: Remove info from internal list on finish
For the same reason that we removed the viewer code, immediately
remove the info from the managers list upon finishing. The
progress info manager at this point is responsible for relaying new
info signals, and maintaining a list if a new window is opened
(thus creating a new progress indicator).