files-view: Update actions immediately on selection change
In the network view, if the row is not selected, clicking the unmount
button does nothing. Indeed, the action is still disabled, despite an
elligible item being selected already.

This happens because the actions state is updated on idle along with
the context menus. This is reasonable as the context menus update can
be expensive.

However, not updating actions state imimediately is unexpected, so
let's immediately update them when selection changes.