placessidebar: Stop requesting size
Nowadays, we have breakpoint-based layouts which take care of ensuring
the sidebar has the desired size.

But the sidebar was itself requesting a size. It was smaller than what
the window layout enforces, so it was not noticed until now that it is
forcing the progress indicator in the sidebar to be clipped off-window.

Do the right and simple thing to fix it.