window-slot: Always hide search bar when changing location
We have code which tries not to hide the search entry in case the user
opens it and, before inserting any query, decides to switch location.
This was introduced by commit 9ab3c5c7d7f7162193559ad643707dc1a2bc69fc

According to that commit message, it was restoring an old behavior that
was lost during a refactoring. I did not confirm this nor track the
original reason for this behavior, but, if I had to guess, I'd say it
comes from before "just-type" search was introduced. Otherwise, why go
to the trouble of keeping the search entry visible when you can just
type anyway?

I see no reason to assume the location change means they want to search
another location. They may have just recalled the exact path they were
looking for, so they lost the reason to search.

Anyway, this behavior is not working at all in recent versions. Not
sure when it stopped working, but apparently nobody has missed it.

Least this dead code interfere with the upcoming global search mode,
let's remove it.