query-editor: Create query even if location is NULL
We have been hitting warnings about NULL queries when doing
global searches, and this was the cause.

There is no good reason not to create a query, even if we are
not going to use it. Also, this is getting in the way of the
implementation of the Search Everywhere place.