query-editor: Create query even if location is NULL
We have been hitting warnings about NULL queries when doing
global searches, and this was the cause.

There is no good reason not to create a query, even if we are
not going to use it. Fix this before we make global search
more prominent.


(cherry picked from commit a5048afb1bc936df12c1ddf6e06261c67a4ce21a)