general: Introduce Search Everywhere place
The current design pattern for global search includes a dedicated view.

For us to adopt this pattern, and to have a major feature in the corner
of the sidebar, introduce a Search Everywhere place, which allows one
to start a global search directly, instead of having to make a regular
search and the upgrade it to global, as was make possible last release.

This place always has a visible search entry and doesn't have a button.
To make this possible, have the slot keep track of whether it is the
Search Everywhere place, for the window to tell the toolbar whether or
not to show the and doesn't have a folder.

Also use this place when launched from the shell search provider.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3013