toolbar: Reserve folder search button space
When the folder search button is hidden (in Search Everywhere place),
the searchbar moves to recenter in the headerbar, because a hidden
widget doesn't allocate space.

This repositioning of the search entry is obviously undesirable.

As a workaround, don't toggle GtkWidget:visible but, instead, use
a stack to swap the button for an empty box to act as its shadow.
The homogenous stack ensure the button alocated space is preserved.