toolbar: Reserve space for hidden buttons
In global search mode, history and folder search buttons are hidden.

This causes the searchbar moves to expand its width because a hidden
widget doesn't allocate space.

This resizing of the search entry is obviously undesirable.

As a workaround, don't toggle GtkWidget:visible but, instead, use
a stack to swap the buttons for an empty box to act as their shadow.
The homogenous stack ensure the button alocated space is preserved.